make 64-bit clean

+++ Examples/HeterogeneousFilters/file_util.h
@@ -47,10 +47,10 @@ inline int open_file(const char* filename, int flags, mode_t mode, int errprint 
    return fd;
 }
 
-inline char* read_file(const char* filename, unsigned& filelen)
+inline char* read_file(const char* filename, unsigned long& filelen)
 {
    ifstream f;
-   int len;
+   long int len;
    char* contents = NULL;
    filelen = 0;
    f.open(filename);
